+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 1.0.0a1 (unreleased)
 
+### Added
+
+- Added tags for adapter amb_event (inLanguage, creator:name, dateCreated, datePublished, dateModified)
+
 ### Fixed
 - The control panel is not unregistered when the package is uninstalled.